Added base class for entities
The Entity class functions as an abstract base class for DDD entity types.

DDD entity type are identified by their identifier and their identifier alone. If two instances of the same entity class, have identifiers with the same value, then they are considered the same even of their contained data differ.

The Equals methods compare the identifiers. Same goes for the equals and not equals operators.

The entity identifier property is marked as protected to make the entity as locked down as possible. If there's a wish to expose the identifier to the outside, then that's the responsibility of the child class.