Added base class for values
The Value class functions as an abstract base class for DDD value types.

DDD value types are identified by their data. If two instances of the same value class contains the same data, they are considered to be equal.

Child classes of the Value type constructs a list of objects that should be included in equality checks. This list is provided to the base class by implementing the abstract method GetObjectsForEqualityCheck.